refactor: Letter, Shape 디렉터리를 이동한다.
