+++ src/main/java/type/Letter.java
@@ -1,4 +1,4 @@
-package domain;
+package type;
 
 public enum Letter {
     TWO("2", 2),
