+++ src/main/java/type/Shape.java
@@ -1,4 +1,4 @@
-package domain;
+package type;
 
 public enum Shape {
