+++ src/test/java/domain/CardDeckTest.java
@@ -2,6 +2,8 @@ package domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import type.Letter;
+import type.Shape;
 
 import java.util.List;
