+++ src/test/java/domain/CardTest.java
@@ -2,6 +2,8 @@ package domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import type.Letter;
+import type.Shape;
 
 import static org.assertj.core.api.Assertions.assertThat;
