+++ src/test/java/domain/DealerTest.java
@@ -3,6 +3,8 @@ package domain;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import type.Letter;
+import type.Shape;
 
 import java.util.ArrayList;
 import java.util.List;
